+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- DepartmentHeader component to wrap Carousel and MainCategories components.
+- Department page template.
+
 ## [1.9.4] - 2018-08-02
 ### Changed
 - Bump `vtex.styleguide` major version.